Fix middleware active user
