+++ website/website/settings.py
@@ -52,11 +52,8 @@ MIDDLEWARE = [
     'django.contrib.auth.middleware.AuthenticationMiddleware',
     'django.contrib.messages.middleware.MessageMiddleware',
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
-]
-
-MIDDLEWARE_CLASSES = (
     'middleware.activeuser_middleware.ActiveUserMiddleware',
-)
+]
 
 CACHES = {
     'default': {
